Moved ApiCall implementation to static EVMBuilder
